Move TextFormat in the context.
