+++ api/src/main/java/openconsensus/tags/propagation/TextFormat.java
@@ -16,8 +16,6 @@
 
 package openconsensus.tags.propagation;
 
-import java.util.List;
-import javax.annotation.Nullable;
 import openconsensus.tags.TagMap;
 
 /**
@@ -34,7 +32,7 @@ import openconsensus.tags.TagMap;
  * <pre>{@code
  * private static final Tagger tagger = Tags.getTagger();
  * private static final TextFormat textFormat =
- *     Tags.getPropagationComponent().getTextFormat();
+ *     Tags.getTagger().getTextFormat();
  * private static final TextFormat.Setter setter =
  *     new TextFormat.Setter<HttpURLConnection>() {
  *       public void put(HttpURLConnection carrier, String key, String value) {
@@ -43,13 +41,10 @@ import openconsensus.tags.TagMap;
  *     };
  *
  * void makeHttpRequest() {
- *   TagMap tagMap = tagger.emptyBuilder().put(K, V).build();
- *   try (Scope s = tagger.withTagMap(tagMap)) {
- *     HttpURLConnection connection =
- *         (HttpURLConnection) new URL("http://myserver").openConnection();
- *     textFormat.inject(tagMap, connection, httpURLConnectionSetter);
- *     // Send the request, wait for response and maybe set the status if not ok.
- *   }
+ *   HttpURLConnection connection =
+ *       (HttpURLConnection) new URL("http://myserver").openConnection();
+ *   textFormat.inject(tagger.getCurrentTagMap(), connection, httpURLConnectionSetter);
+ *   // Send the request, wait for response and maybe set the status if not ok.
  * }
  * }</pre>
  *
@@ -58,7 +53,7 @@ import openconsensus.tags.TagMap;
  * <pre>{@code
  * private static final Tagger tagger = Tags.getTagger();
  * private static final TextFormat textFormat =
- *     Tags.getPropagationComponent().getTextFormat();
+ *     Tags.getTagger().getTextFormat();
  * private static final TextFormat.Getter<HttpRequest> getter = ...;
  *
  * void onRequestReceived(HttpRequest request) {
@@ -71,90 +66,4 @@ import openconsensus.tags.TagMap;
  *
  * @since 0.1.0
  */
-public abstract class TextFormat {
-
-  /**
-   * The propagation fields defined. If your carrier is reused, you should delete the fields here
-   * before calling {@link #inject(TagMap, Object, Setter)}.
-   *
-   * <p>For example, if the carrier is a single-use or immutable request object, you don't need to
-   * clear fields as they couldn't have been set before. If it is a mutable, retryable object,
-   * successive calls should clear these fields first.
-   *
-   * @since 0.1.0
-   */
-  // The use cases of this are:
-  // * allow pre-allocation of fields, especially in systems like gRPC Metadata
-  // * allow a single-pass over an iterator (ex OpenTracing has no getter in TextMap)
-  public abstract List<String> fields();
-
-  /**
-   * Injects the tag context downstream. For example, as http headers.
-   *
-   * @param tagMap the tag context.
-   * @param carrier holds propagation fields. For example, an outgoing message or http request.
-   * @param setter invoked for each propagation key to add or remove.
-   * @throws SerializationException if the given tag context cannot be serialized.
-   * @since 0.1.0
-   */
-  public abstract <C> void inject(TagMap tagMap, C carrier, Setter<C> setter)
-      throws SerializationException;
-
-  /**
-   * Class that allows a {@code TextFormat} to set propagated fields into a carrier.
-   *
-   * <p>{@code Setter} is stateless and allows to be saved as a constant to avoid runtime
-   * allocations.
-   *
-   * @param <C> carrier of propagation fields, such as an http request
-   * @since 0.1.0
-   */
-  public abstract static class Setter<C> {
-
-    /**
-     * Replaces a propagated field with the given value.
-     *
-     * <p>For example, a setter for an {@link java.net.HttpURLConnection} would be the method
-     * reference {@link java.net.HttpURLConnection#addRequestProperty(String, String)}
-     *
-     * @param carrier holds propagation fields. For example, an outgoing message or http request.
-     * @param key the key of the field.
-     * @param value the value of the field.
-     * @since 0.1.0
-     */
-    public abstract void put(C carrier, String key, String value);
-  }
-
-  /**
-   * Extracts the tag context from upstream. For example, as http headers.
-   *
-   * @param carrier holds propagation fields. For example, an outgoing message or http request.
-   * @param getter invoked for each propagation key to get.
-   * @throws DeserializationException if the input is invalid
-   * @since 0.1.0
-   */
-  public abstract <C> TagMap extract(C carrier, Getter<C> getter) throws DeserializationException;
-
-  /**
-   * Class that allows a {@code TextFormat} to read propagated fields from a carrier.
-   *
-   * <p>{@code Getter} is stateless and allows to be saved as a constant to avoid runtime
-   * allocations.
-   *
-   * @param <C> carrier of propagation fields, such as an http request
-   * @since 0.1.0
-   */
-  public abstract static class Getter<C> {
-
-    /**
-     * Returns the first value of the given propagation {@code key} or returns {@code null}.
-     *
-     * @param carrier carrier of propagation fields, such as an http request
-     * @param key the key of the field.
-     * @return the first value of the given propagation {@code key} or returns {@code null}.
-     * @since 0.1.0
-     */
-    @Nullable
-    public abstract String get(C carrier, String key);
-  }
-}
+public abstract class TextFormat extends openconsensus.context.propagation.TextFormat<TagMap> {}
